Reduce the amount of html code generated with default settings
Earlier some of the html attributes were specified explicitly,
despite being set to 0 or 100%. Now under such conditions these
attributes are skipped. Also the conditions in the get_margins
method are sorted in ascending order based on the number of symbols
they return.
The align attribute is now in the div tag, and if some specific
paragraph's alignment differs, the align attribute is added directly
to its p tag.

Minor changes:
* In classes' names replaced 'File' with 'Document'
* Removed staticmethod decorator from the methods of class Format
* Added helper functions to get width and indent from the settings
* Now two files are created for each converted document (.txt + .html)